+++ api/hydrophoneApi.go
@@ -265,12 +265,12 @@ func (a *Api) checkFoundConfirmations(token string, res http.ResponseWriter, res
 				//report and move on
 				log.Println("Error getting profile", err.Error())
 			}
-			if results[i].Team != nil && results[i].Team.TeamID != "" {
-				team, err := a.perms.GetTeam(token, results[i].Team.TeamID)
+			if results[i].Team != nil && results[i].Team.ID != "" {
+				team, err := a.perms.GetTeam(token, results[i].Team.ID)
 				if err != nil {
 					log.Println("Error getting team", err.Error())
 				} else {
-					results[i].Team.TeamName = team.Name
+					results[i].Team.Name = team.Name
 				}
 			}
 		}