+++ api/signup.go
@@ -31,6 +31,7 @@ const (
 	STATUS_INVALID_BIRTHDAY  = "Birthday specified is invalid"
 	STATUS_MISMATCH_BIRTHDAY = "Birthday specified does not match patient birthday"
 	STATUS_PATIENT_NOT_AUTH  = "Patient cannot be member of care team"
+	STATUS_MEMBER_NOT_AUTH   = "Non patient users cannot be a patient of care team"
 )
 
 const (