+++ CHANGELOG.md
@@ -3,14 +3,15 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
-## unreleased 
+## 1.6.2 
 ### Changed
 - YLP-682 Add team information in received invitations
 
 ### Fixed 
-- YLP-709 Bug: get the coorect ID to dismiss an invite
-- YLP-708 A HCP user shouldn't invite a patient to join a team 
+- YLP-709 Bug: get the correct ID to dismiss an invite
+- YLP-708 A HCP user shouldn't invite a patient to join a team and vice-versa
 - YLP-743 A team invitation for a patient is valid if the patient account exists
+
 ## 1.6.1 - 2021-05-04
 ### Changed
 - YLP-710 Modify route /accept/team/invite to acknowledge notifications that does not requires a specific action (i.e. change member role)